+++ Samples/Office/ReadOfficeMetadata/Program.cs
@@ -26,12 +26,9 @@ namespace ReadOfficeMetadata
             // Initialize PDF support. Initialize for each process or thread.
             ImGearPDF.Initialize();
 
-            // Initialize Office support for each process or thread by providing a path to your LibreOffice installation.
-            // See ImageGear documentation for more information.
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
-                ImGearOffice.Initialize(@"path/to/linux/libreoffice/bin/directory");
-            else
-                ImGearOffice.Initialize(@"path\to\windows\libreoffice\bin\directory");
+            // Initialize Office support for each process or thread. You must supply the path to your LibreOffice installation
+            // via the IMAGEGEAR_LIBREOFFICE_PATH environment variable. See ImageGear documentation for more information.
+            ImGearOffice.Initialize();
 
             // Load image page.
             ImGearDocument imGearDocument;