+++ Samples/PDF/PDFContentExtractText/Program.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.IO;
-using ImageGear.Evaluation;
 using ImageGear.Formats;
 using ImageGear.Formats.PDF;